Improve BY adjustment performance by approximating harmonic numbers
The Benjamini-Yekutieli uses the harmonic number in its step function. The
previous implementation through summation of the individual 1/x terms was very
inefficient for large x. This has been replaced by an approximation of the
harmonic number with constant run time. In comparison with the previous
implementation,
- run time is equivalent for small n
- run time is reduced by several orders of magnitude for large n
- the approximated values are within one order of magnitude to floating point precision
compared to the exact numbers, similarly to the precision obtained through summation.