Turn `PValues` type immutable and persistent
This change modifies the definition and construction of the `PValues` type, in order to achieve a persistent representation of the data. Since the actual p-value vector in the type are mutable, even with an immutable `PValues` type, changes to the original object representing the p-values affect the validity and consistency of the `PValues` object:

```julia
p = collect(0.1:0.1:0.9)
pv = PValues(p)
p[1] = -10.0
minimum(p) # => 0.1, not updated
```

As a consequence,
a) the extrema stored in the object may not match the extrema of the actual values anymore
b) the object can become invalid without noticing, although the validity was checked at the initial construction.

In the new version, the p-value vector is copied during construction which address these issues. This comes at the cost of copying the vector once during construction, but it appears neglectable in the light of either having to deal with potentially invalid p-values.